Install Debian's rustc package instead of the devcontainer rust feature
Just installing rustc is a lighter weight way of installing Rust without compiling, and does not introduce the rust-related vscode extensions installed by the feature which are unecessary for most ruby development. This gives us rust version 1.63, which is acceptable for now. If we need a specific version in the future we may need to compile rust ourselves.