Explicitly set the user to vscode in the Ruby image Dockerfile
Since the metadata from devcontainer.json is not present in the image, other Dev Container implementations will not set this user (which is default for vscode and created by the debian base image). So lets explicitly set it so that those Dev Container implementations use the right user when using this image.